+++ packages/clerk-js/src/ui/components/UserProfile/UserProfileNavbar.tsx
@@ -2,6 +2,7 @@ import React from 'react';
 
 import { NavBar, NavbarContextProvider } from '@/ui/elements/Navbar';
 
+import { USER_PROFILE_NAVBAR_ROUTE_ID } from '../../constants';
 import { useUserProfileContext } from '../../contexts';
 import { localizationKeys } from '../../localization';
 import type { PropsOfComponent } from '../../styledSystem';
@@ -9,14 +10,16 @@ import type { PropsOfComponent } from '../../styledSystem';
 export const UserProfileNavbar = (
   props: React.PropsWithChildren<Pick<PropsOfComponent<typeof NavBar>, 'contentRef'>>,
 ) => {
-  const { pages } = useUserProfileContext();
+  const { pages, apiKeysProps } = useUserProfileContext();
+
+  const routes = pages.routes.filter(r => r.id !== USER_PROFILE_NAVBAR_ROUTE_ID.API_KEYS || !apiKeysProps?.hide);
 
   return (
     <NavbarContextProvider contentRef={props.contentRef}>
       <NavBar
         title={localizationKeys('userProfile.navbar.title')}
         description={localizationKeys('userProfile.navbar.description')}
-        routes={pages.routes}
+        routes={routes}
         contentRef={props.contentRef}
       />
       {props.children}